src/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cli-exit-tools==1.2.3.2
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
jira==3.5.0
lib-detect-testenv==2.0.3
oauthlib==3.2.2
packaging==23.0
pathlib3x==2.0.2.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
smmap==5.0.0
typing_extensions==4.5.0
urllib3==1.26.15